+++ HtmlRun.Common/Models/CallModel.cs
@@ -11,5 +11,10 @@ public class CallModel
   public List<CallArgumentModel> Arguments { get; set; } = new List<CallArgumentModel>();
 
   public bool IsSpecial { get; set; } = false;
+
+  public override string ToString()
+  {
+    return $"#{this.Index}# {FunctionName} ({string.Join(", ", this.Arguments.Select(m => m.ToString()))})";
+  }
 }