+++ HtmlRun.Runtime/Constants/MathInstructionsSet.cs
@@ -1,0 +1,6 @@
+﻿namespace HtmlRun.Runtime.Constants;
+
+public static class MathInstructionsSet
+{
+  public const string Clamp = "Clamp";
+}