+++ HtmlRun.Runtime/Providers/JumpStatementsProvider.cs
@@ -97,11 +97,11 @@ class PopArgumentsCmd : INativeInstruction
     {
       return ctx =>
       {
-        var args = ctx.PopArgumentsAndValues() ?? throw new NullReferenceException("No arguments to pop.");
-        
-        if (args.Arguments == null)
+        var args = ctx.PopArgumentsAndValues();
+
+        if (args?.Arguments == null)
         {
-          throw new InvalidOperationException($"Error trying to read arguments of group {args.Label}.");
+          return;
         }
 
         foreach (var arg in args.Arguments)